Switches back to `governor` implementation.
Given the way PRs are progressing, it looks like we'll be able to use
`governor` as the single rate limiting library.
