+++ Cargo.toml
@@ -19,3 +19,13 @@ default = ["event_loop"]
 
 event_loop = ["nginee_event_loop"]
 rate_limit = ["nginee_event_loop/rate_limit"]
+
+wasm-bindgen = ["nginee_event_loop/wasm-bindgen"]
+
+[patch.crates-io]
+# https://github.com/antifuchs/governor/pull/35
+governor = { git = "https://github.com/azriel91/governor.git", rev = "4a34ce9" }
+# https://github.com/sebcrozet/instant/pull/17
+instant = { git = "https://github.com/sebcrozet/instant.git", rev = "ef11b02" }
+# https://github.com/Amanieu/parking_lot/pull/231
+parking_lot = { git = "https://github.com/azriel91/parking_lot.git", rev = "37ceeb3" }
