+++ crate/event_loop/Cargo.toml
@@ -10,13 +10,10 @@ test = false
 
 [dependencies]
 futures = { version =  "0.3.5", default-features = false }
-
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-governor = { version = "0.2.0", optional = true }
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-async-std = { version = "1.6.0-beta.1", optional = true }
+governor = { version = "0.2.0", default-features = false, features = ["std"], optional = true }
+parking_lot = { version = "0.10.2", optional = true }
 
 [features]
 default = []
-rate_limit = ["async-std", "governor"]
+rate_limit = ["governor"]
+wasm-bindgen = ["governor/wasm-bindgen", "parking_lot/wasm-bindgen"]
