+++ crate/workspace_tests/Cargo.toml
@@ -4,12 +4,10 @@ version = "0.1.0"
 authors = ["Azriel Hoh <azriel91@gmail.com>"]
 edition = "2018"
 
-[target.'cfg(not(target_arch = "wasm32"))'.dev-dependencies]
-smol = "0.1.4"
-
 [dev-dependencies]
 crossbeam = "0.7.3"
 nginee = { path = "../..", features = ["event_loop"] }
+smol = "0.1.4"
 
 [features]
 default = ["rate_limit"]
